requirements.txt
anyio==4.6.2.post1
av==13.1.0
bchlib==2.1.3
bitmath==1.3.3.1
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.2.0
deprecation==2.1.0
dghs-imgutils==0.7.0
emoji==2.11.1
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fsspec==2024.10.0
h11==0.14.0
h2==4.1.0
hbutils==0.10.1
hfutils==0.5.1
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
mpmath==1.3.0
natsort==8.4.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
onnxruntime-gpu==1.20.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
piexif==1.1.3
pillow==11.0.0
pilmoji==2.0.4
pixivpy3==3.7.5
protobuf==5.28.3
pyclipper==1.3.0.post6
pyparsing==3.2.0
pyquery==2.0.1
pyrate-limiter==3.7.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
random-user-agent==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
URLObject==2.4.3
waifuc @ git+https://github.com/deepghs/waifuc.git@efe5c49171a94a6441a79b7d7b595d4b3a4eb51f
xmltodict==0.14.2
